# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Legacy Doc API",
    "description": "Documentacao automatica multi-linguagem com orquestracao multi-agente e multi-provedor.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/v1/jobs": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Create Job",
        "description": "Enqueue a job and return 202 immediately.",
        "operationId": "create_job_v1_jobs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/RepositoryJobRequest"
                  },
                  {
                    "$ref": "#/components/schemas/SnippetJobRequest"
                  }
                ],
                "discriminator": {
                  "propertyName": "job_type",
                  "mapping": {
                    "document_repository": "#/components/schemas/RepositoryJobRequest",
                    "document_snippet": "#/components/schemas/SnippetJobRequest"
                  }
                },
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_v1_jobs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/upload": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Create Upload Job",
        "description": "Enqueue documentation for an uploaded archive.",
        "operationId": "create_upload_job_v1_jobs_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_upload_job_v1_jobs_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_v1_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "Cancel",
        "description": "Cancel a job that has not started.",
        "operationId": "cancel_v1_jobs__job_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_upload_job_v1_jobs_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "Arquivo .zip com o codigo-fonte"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "output_language": {
            "type": "string",
            "title": "Output Language",
            "default": "pt-BR"
          },
          "depth": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GenerationDepth"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_create_upload_job_v1_jobs_upload_post"
      },
...
      "GenerationDepth": {
        "type": "string",
        "enum": [
          "basic",
          "standard",
          "pro"
        ],
        "title": "GenerationDepth",
        "description": "How much analysis a single job asks for."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "JobListResponse"
      },
      "JobResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "job_type": {
            "type": "string",
            "title": "Job Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress_percent": {
            "type": "integer",
            "title": "Progress Percent"
          },
          "progress_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress Message"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "document_count": {
            "type": "integer",
            "title": "Document Count",
            "default": 0
          },
          "depth": {
            "type": "string",
            "title": "Depth",
            "description": "Profundidade efetivamente aplicada a este job."
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "description": "O que foi analisado: a URL do repositorio ou o nome do arquivo. Sem isto, uma lista de jobs so mostra datas e a pessoa nao reconhece qual e qual."
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_type",
          "status",
          "progress_percent",
          "progress_message",
          "attempts",
          "project_id",
          "created_at",
          "started_at",
          "finished_at",
          "error_code",
          "error_message",
          "depth"
        ],
        "title": "JobResponse"
      },
...
      "RepositoryJobRequest": {
        "properties": {
          "job_type": {
            "type": "string",
            "const": "document_repository",
            "title": "Job Type",
            "default": "document_repository"
          },
          "repo_url": {
            "type": "string",
            "title": "Repo Url"
          },
          "branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch"
          },
          "paths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Paths",
            "description": "Caminhos a documentar. Vazio processa o repositorio inteiro ate o limite do plano."
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "output_language": {
            "type": "string",
            "title": "Output Language",
            "default": "pt-BR"
          },
          "depth": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GenerationDepth"
              },
              {
                "type": "null"
              }
            ],
            "description": "Profundidade da analise. Omitido usa o maximo do plano. Pedir acima do plano nao da erro: e reduzido ao teto, e a resposta informa o que foi efetivamente usado."
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url"
          }
        },
        "type": "object",
        "required": [
          "repo_url"
        ],
        "title": "RepositoryJobRequest",
        "description": "Document files from a remote repository."
      },
      "SnippetJobRequest": {
        "properties": {
          "job_type": {
            "type": "string",
            "const": "document_snippet",
            "title": "Job Type",
            "default": "document_snippet"
          },
          "path": {
            "type": "string",
            "title": "Path",
            "description": "Caminho do arquivo, usado para detectar a linguagem"
          },
          "content": {
            "type": "string",
            "maxLength": 1000000,
            "minLength": 1,
            "title": "Content"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "output_language": {
            "type": "string",
            "title": "Output Language",
            "default": "pt-BR"
          },
          "depth": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GenerationDepth"
              },
              {
                "type": "null"
              }
            ],
            "description": "Profundidade da analise. Omitido usa o maximo do plano. Pedir acima do plano nao da erro: e reduzido ao teto, e a resposta informa o que foi efetivamente usado."
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url"
          }
        },
        "type": "object",
        "required": [
          "path",
          "content"
        ],
        "title": "SnippetJobRequest",
        "description": "Document inline code. Used by the VS Code extension."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}